Evaluation article code: Shallow vs Deep copy
