+++ core-java-modules/core-java-object-copy/src/test/java/com/baeldung/deepcopy/DeepCopyUnitTest.java
@@ -1,13 +1,13 @@
-package com.baeldung.shallowcopy;
+package com.baeldung.deepcopy;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
 
-public class ShallowCopyTest {
+public class DeepCopyUnitTest {
 
     @Test
-    public void whenModifyingClonedPersonFields_thenOriginalPersonAddressChanges() {
+    public void whenModifyingCopiedPerson_thenOriginalDoesNotChange() {
         Address address = new Address("123 Main Street", "New York", "NY 10030");
         Person firstPerson = new Person("John Smith", 26, address);
         Person secondPerson = firstPerson.clone();
@@ -19,14 +19,14 @@ public class ShallowCopyTest {
 
         assertThat(firstPerson).isNotSameAs(secondPerson);
         assertThat(firstPerson.getAddress()
-            .getAddressLine()).isSameAs(secondPerson.getAddress()
+            .getAddressLine()).isNotSameAs(secondPerson.getAddress()
             .getAddressLine());
         assertThat(firstPerson.getAge()).isNotEqualTo(secondPerson.getAge());
         assertThat(firstPerson.getName()).isNotEqualTo(secondPerson.getName());
     }
 
     @Test
-    public void whenModifyingConstructorCopiedPersonFields_thenOriginalPersonAddressChanges() {
+    public void whenModifyingConstructorCopiedPerson_thenOriginalPersonDoesNotChange() {
         Address address = new Address("123 Main Street", "New York", "NY 10030");
         Person firstPerson = new Person("John Smith", 26, address);
         Person secondPerson = new Person(firstPerson);
@@ -38,9 +38,10 @@ public class ShallowCopyTest {
 
         assertThat(firstPerson).isNotSameAs(secondPerson);
         assertThat(firstPerson.getAddress()
-            .getAddressLine()).isSameAs(secondPerson.getAddress()
+            .getAddressLine()).isNotSameAs(secondPerson.getAddress()
             .getAddressLine());
         assertThat(firstPerson.getAge()).isNotEqualTo(secondPerson.getAge());
         assertThat(firstPerson.getName()).isNotEqualTo(secondPerson.getName());
     }
+
 }
